requirements.txt
pyTelegramBotAPI==4.29.1
deep-translator==1.11.4
python-dotenv==1.0.1
